Extract BlobUploader class from PdfGenerator class
