+++ PdfContent.cs
@@ -1,0 +1,10 @@
+using System.IO;
+using Newtonsoft.Json;
+
+namespace az_function
+{
+    public class PdfContent
+    {
+        public byte[] Content { set; get; }
+    }
+}
