+++ PdfGenerator.cs
@@ -16,30 +16,24 @@ namespace az_function
     public static class PdfGenerator
     {
         [FunctionName("Chaining_GeneratePdf")]
-        public static async Task<IActionResult> GeneratePdf([ActivityTrigger] string request, ILogger log)
+        public static async Task<byte[]> GeneratePdf([ActivityTrigger] string request, ILogger log)
         {
             log.LogInformation("C# HTTP trigger function processed a request.");
+
+            if (GlobalFontSettings.FontResolver == null)
+            {
+                // Set the global font resolver
+                GlobalFontSettings.FontResolver = new CustomFontResolver();
+            }
 
             // Generate PDF from the input string
             byte[] pdfBytes = GeneratePdf(request);
 
-            // Upload the PDF to Azure Blob Storage
-            string connectionString = "DefaultEndpointsProtocol=https;AccountName=tobiasodionia;AccountKey=I6LkSh1JAnsyDFQmzfYekc6vP07/if7XVZAKKJQ0Mojjhwcusas3AEN30XV9muojtPRkasfCQwva+AStUQjZeg==;EndpointSuffix=core.windows.net";
-            string containerName = "ia-pdf";
-            string blobName = Guid.NewGuid().ToString() + ".pdf";
-            string mimeType = MimeMapping.MimeUtility.GetMimeMapping(blobName);
-
-            var blobUrl = await UploadToBlobStorage(connectionString, containerName, blobName, pdfBytes, mimeType);
-
-            // Return the URL of the stored PDF
-            return new OkObjectResult($"PDF stored at: {blobUrl}");
+            return pdfBytes;
         }
 
         private static byte[] GeneratePdf(string content)
         {
-            // Set the global font resolver
-            GlobalFontSettings.FontResolver = new CustomFontResolver();
-
             using (MemoryStream stream = new MemoryStream())
             {
                 using (PdfDocument pdfDocument = new PdfDocument())
@@ -60,7 +54,7 @@ namespace az_function
 
         private static void DrawWrappedText(XGraphics gfx, XFont font, string text, XRect rect)
         {
-           // Split the text into words
+            // Split the text into words
             string[] words = text.Split(' ');
 
             // Initialize variables
@@ -92,28 +86,6 @@ namespace az_function
 
             // Draw the remaining text
             gfx.DrawString(currentLine.Trim(), font, XBrushes.Black, rect, XStringFormats.TopLeft);
-        }
-
-        private static async Task<string> UploadToBlobStorage(string connectionString, string containerName, string blobName, byte[] content, string fileMimeType)
-        {
-            var storageAccount = CloudStorageAccount.Parse(connectionString);
-            var blobClient = storageAccount.CreateCloudBlobClient();
-            var container = blobClient.GetContainerReference(containerName);
-
-            if (await container.CreateIfNotExistsAsync())
-            {
-                await container.SetPermissionsAsync(new BlobContainerPermissions { PublicAccess = BlobContainerPublicAccessType.Blob });
-            }
-
-            if (blobName != null && content != null)
-            {
-                var blob = container.GetBlockBlobReference(blobName);
-                blob.Properties.ContentType = fileMimeType;
-                await blob.UploadFromByteArrayAsync(content, 0, content.Length);
-                return blob.Uri.AbsoluteUri;
-            }
-
-            return "";
         }
     }
 }
